refactor GhostMapParser to be in utility package
