+++ jpacman/src/test/java/utility/GhostMapParser.java
@@ -1,4 +1,4 @@
-package nl.tudelft.jpacman.npc.ghost;
+package utility;
 
 import java.util.List;
 import nl.tudelft.jpacman.board.BoardFactory;
@@ -6,6 +6,7 @@ import nl.tudelft.jpacman.board.Square;
 import nl.tudelft.jpacman.level.LevelFactory;
 import nl.tudelft.jpacman.level.MapParser;
 import nl.tudelft.jpacman.npc.Ghost;
+import nl.tudelft.jpacman.npc.ghost.GhostFactory;
 
 /**
  * A test helper utility for writing the ghost unit tests.
